+++ sdk/communication/communication-call-automation/src/callAutomationEventParser.ts
@@ -26,8 +26,8 @@ import {
   ContinuousDtmfRecognitionToneReceived,
   ContinuousDtmfRecognitionToneFailed,
   ContinuousDtmfRecognitionStopped,
-  SendDtmfCompleted,
-  SendDtmfFailed,
+  SendDtmfTonesCompleted,
+  SendDtmfTonesFailed,
 } from "./models/events";
 
 import { CloudEventMapper } from "./models/mapper";
@@ -120,11 +120,11 @@ export function parseCallAutomationEvent(
         kind: "ContinuousDtmfRecognitionStopped",
       } as ContinuousDtmfRecognitionStopped;
       break;
-    case "Microsoft.Communication.SendDtmfCompleted":
-      callbackEvent = { kind: "SendDtmfCompleted" } as SendDtmfCompleted;
+    case "Microsoft.Communication.SendDtmfTonesCompleted":
+      callbackEvent = { kind: "SendDtmfTonesCompleted" } as SendDtmfTonesCompleted;
       break;
-    case "Microsoft.Communication.SendDtmfFailed":
-      callbackEvent = { kind: "SendDtmfFailed" } as SendDtmfFailed;
+    case "Microsoft.Communication.SendDtmfTonesFailed":
+      callbackEvent = { kind: "SendDtmfTonesFailed" } as SendDtmfTonesFailed;
       break;
     default:
       throw new TypeError(`Unknown Call Automation Event type: ${eventType}`);