+++ sdk/communication/communication-call-automation/src/models/models.ts
@@ -55,8 +55,8 @@ export interface CallLocator {
   kind: CallLocatorType;
 }
 
-/** Defines values for Gender that the service accepts. */
-export enum Gender {
+/** Defines values for VoiceKind that the service accepts. */
+export enum VoiceKind {
   /** Male */
   Male = "male",
   /** Female */
@@ -78,7 +78,7 @@ export interface FileSource extends PlaySource {
 export interface TextSource extends PlaySource {
   text: string;
   sourceLocale?: string;
-  voiceGender?: Gender;
+  voiceKind?: VoiceKind;
   voiceName?: string;
   customVoiceEndpointId?: string;
   readonly kind: "textSource";
@@ -127,8 +127,8 @@ export enum DtmfTone {
   Asterisk = "asterisk",
 }
 
-/** A Recognize Choice */
-export interface Choice {
+/** A Recognition Choice */
+export interface RecognitionChoice {
   /** Identifier for a given choice */
   label: string;
   /** List of phrases to recognize */