+++ sdk/communication/communication-call-automation/src/models/responses.ts
@@ -57,8 +57,8 @@ export interface RemoveParticipantResult {
   operationContext?: string;
 }
 
-/** The response payload for muting participants from the call. */
-export interface MuteParticipantsResult {
+/** The response payload for muting participant from the call. */
+export interface MuteParticipantResult {
   /** The operation context provided by client. */
   operationContext?: string;
 }
@@ -70,7 +70,7 @@ export interface RecordingStateResult {
 }
 
 /** The response payload for sending DTMF tones. */
-export interface SendDtmfResult {
+export interface SendDtmfTonesResult {
   /** The operation context provided by client. */
   operationContext?: string;
 }